// src/types/api/service/models/ClipBookDto.ts
export interface ClipBookDto {
    /**
     * 
     * @type {number}
     * @memberof ClipBookDto
     */
    clipBookId?: number;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    shareCode?: string;
    /**
     * 
     * @type {number}
     * @memberof ClipBookDto
     */
    companyId?: number;
    /**
     * 
     * @type {number}
     * @memberof ClipBookDto
     */
    groupId?: number;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    startYear?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    startMonth?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    startDay?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    endYear?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    endMonth?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    endDay?: string;
    /**
     * 뉴스 목록
     * @type {Array<number>}
     * @memberof ClipBookDto
     */
    newslist?: Array<number>;
    /**
     * 뉴스 검색을 위한 인덱스 목록(YYYYMM)
     * @type {Array<string>}
     * @memberof ClipBookDto
     */
    newsIndexlist?: Array<string>;
    /**
     * 보도자료 목록
     * @type {Array<number>}
     * @memberof ClipBookDto
     */
    prIdlist?: Array<number>;
    /**
     * 보도자료 목록
     * @type {Array<IdTitleDto>}
     * @memberof ClipBookDto
     */
    prlist?: Array<IdTitleDto>;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    regisAt?: string;
    /**
     * 
     * @type {string}
     * @memberof ClipBookDto
     */
    updateAt?: string;
    /**
     * 
     * @type {UserDtoForSimple}
     * @memberof ClipBookDto
     */
    owner?: UserDtoForSimple;
    /**
     * 
     * @type {UserDtoForSimple}
     * @memberof ClipBookDto
     */
    register?: UserDtoForSimple;
    /**
     * 
     * @type {UserDtoForSimple}
     * @memberof ClipBookDto
     */
    updater?: UserDtoForSimple;
}