// src/types/api/service/models/PageClipBookDto.ts
export interface PageClipBookDto {
    /**
     * 
     * @type {number}
     * @memberof PageClipBookDto
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PageClipBookDto
     */
    totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageClipBookDto
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageClipBookDto
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageClipBookDto
     */
    size?: number;
    /**
     * 
     * @type {Array<ClipBookDto>}
     * @memberof PageClipBookDto
     */
    content?: Array<ClipBookDto>;
    /**
     * 
     * @type {number}
     * @memberof PageClipBookDto
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageClipBookDto
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageClipBookDto
     */
    numberOfElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageClipBookDto
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageClipBookDto
     */
    empty?: boolean;
}